requirements/base.txt
arrow==0.10.0
beautifulsoup4==4.5.3
DistributedLock==1.2
Django==1.11
django-annoying==0.10.3
django-emoticons==1.2
django-forms-bootstrap==3.1.0
django-instagram-api==0.3.7
django-jquery==3.1.0
django-m2m-history==0.3.6
django-model-utils==3.0.0
django-oauth-tokens==0.6.3
django-paypal==0.3.3
django-picklefield==0.3.2
django-social-api==0.1.1
django-taggit==0.22.1
django-tinymce==2.6.0
django-twitter-api==0.2.10
django-youtube-api==0.0.2
google-api-python-client==1.6.2
httplib2==0.10.3
oauth2client==4.0.0
oauthlib==2.0.2
Pillow==2.9.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
python-dateutil==2.6.0
python-instagram==1.3.0
python-memcached==1.58
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
tweepy==3.5.0
uritemplate==3.0.0
wheel==0.24.0
